# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TripMatch API - Group travel planning with swipe-based matching
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups for current user
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new trip group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Not found
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupBody"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a group via invite code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupBody"
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Invalid invite code
  /groups/{id}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveGroupResult"
  /groups/{id}/members:
    get:
      operationId: getGroupMembers
      tags: [groups]
      summary: Get all members of a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupMember"
  /groups/{id}/preferences:
    post:
      operationId: submitPreferences
      tags: [groups]
      summary: Submit user preferences for a trip group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPreferencesBody"
      responses:
        "200":
          description: Preferences saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
  /groups/{id}/results:
    get:
      operationId: getGroupResults
      tags: [matching]
      summary: Get matching results for a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matching results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchingResults"
  /groups/{id}/stats:
    get:
      operationId: getGroupStats
      tags: [groups]
      summary: Get group activity stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupStats"
...
components:
...
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        status:
          type: string
          enum: [pending, swiping, matched, planning]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdByUserId:
          type: string
      required:
        - id
        - name
        - inviteCode
        - status
        - memberCount
        - createdAt
        - createdByUserId
    GroupDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        status:
          type: string
          enum: [pending, swiping, matched, planning]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdByUserId:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
        topDestinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationScore"
      required:
        - id
        - name
        - inviteCode
        - status
        - memberCount
        - createdAt
        - createdByUserId
        - members
        - topDestinations
    GroupMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        groupId:
          type: integer
        role:
          type: string
          enum: [creator, member]
        hasCompletedPreferences:
          type: boolean
        swipeCount:
          type: integer
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - groupId
        - role
        - hasCompletedPreferences
        - swipeCount
        - displayName
        - joinedAt
    CreateGroupBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
    UpdateGroupBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    JoinGroupBody:
      type: object
      properties:
        inviteCode:
          type: string
      required:
        - inviteCode
    LeaveGroupResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    SubmitPreferencesBody:
      type: object
      properties:
        budgetMin:
          type: number
        budgetMax:
          type: number
        travelTypes:
          type: array
          items:
            type: string
        climate:
          type: string
        activityLevel:
          type: string
        availableDates:
          type: ["string", "null"]
      required:
        - budgetMin
        - budgetMax
        - travelTypes
        - climate
        - activityLevel
...
    DestinationScore:
      type: object
      properties:
        destinationId:
          type: integer
        destinationName:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        groupScore:
          type: number
        matchPercentage:
          type: number
        consensusLevel:
          type: string
          enum: [high, medium, low]
        likeCount:
          type: integer
        dislikeCount:
          type: integer
        superlikeCount:
          type: integer
        totalVotes:
          type: integer
      required:
        - destinationId
        - destinationName
        - country
        - imageUrl
        - groupScore
        - matchPercentage
        - consensusLevel
        - likeCount
        - dislikeCount
        - superlikeCount
        - totalVotes
    MatchingResults:
      type: object
      properties:
        groupId:
          type: integer
        totalMembers:
          type: integer
        membersCompleted:
          type: integer
        completionPercent:
          type: number
        topDestinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationScore"
        consensusSummary:
          type: string
        hasEnoughData:
          type: boolean
      required:
        - groupId
        - totalMembers
        - membersCompleted
        - completionPercent
        - topDestinations
        - consensusSummary
        - hasEnoughData
    GroupStats:
      type: object
      properties:
        groupId:
          type: integer
        totalSwipes:
          type: integer
        membersSwipedCount:
          type: integer
        membersPreferencesCount:
          type: integer
        topMatch:
          $ref: "#/components/schemas/DestinationScore"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - groupId
        - totalSwipes
        - membersSwipedCount
        - membersPreferencesCount
        - recentActivity
    ActivityItem:
      type: object
      properties:
        type:
          type: string
          enum: [swipe, join, preference]
        userId:
          type: string
        displayName:
          type: string
        destinationName:
          type: ["string", "null"]
        action:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - type
        - userId
        - displayName
        - action
        - createdAt
...