# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  type: ActivityItemType;
  userId: string;
  displayName: string;
  /** @nullable */
  destinationName?: string | null;
  action: string;
  createdAt: Date;
}