# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGroupBody {
  name: string;
  /** @nullable */
  description?: string | null;
}