# lib/api-zod/src/generated/types/groupDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */
import type { DestinationScore } from "./destinationScore";
import type { GroupDetailStatus } from "./groupDetailStatus";
import type { GroupMember } from "./groupMember";

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  inviteCode: string;
  status: GroupDetailStatus;
  memberCount: number;
  createdAt: Date;
  createdByUserId: string;
  members: GroupMember[];
  topDestinations: DestinationScore[];
}