# lib/api-zod/src/generated/types/groupMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TripMatch API - Group travel planning with swipe-based matching
 * OpenAPI spec version: 0.1.0
 */
import type { GroupMemberRole } from "./groupMemberRole";

export interface GroupMember {
  id: number;
  userId: string;
  groupId: number;
  role: GroupMemberRole;
  hasCompletedPreferences: boolean;
  swipeCount: number;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  joinedAt: Date;
}